test: make jsonWatcher test close file watcher

+++ test/module.chinachu-common.test.js
@@ -1,64 +1,78 @@
 "use strict";
 
-var os     = require('os');
-var fs     = require('fs');
-var should = require('should');
+var os     = require("os");
+var path   = require("path");
+var fs     = require("fs");
+var should = require("should");
 
-var chinachu = require('chinachu-common');
+var chinachu = require("chinachu-common");
 
-var testDataPath = os.tmpDir() + '/chinachu-test-' + new Date().getTime() + '.json';
+var testDataPath = path.join(os.tmpdir(), "chinachu-test-" + Date.now() + ".json");
+var watcher = null;
 
-describe('(init)', function() {
-	
+describe("(init)", function() {
 	var testData = {
 		a: 0,
 		b: 1,
-		c: '',
-		d: 'string',
+		c: "",
+		d: "string",
 		e: null,
 		f: {},
-		g: { a: 0, b: 1, c: '', d: 'string', e: null, f: {}, h: [] },
+		g: { a: 0, b: 1, c: "", d: "string", e: null, f: {}, h: [] },
 		h: [],
-		i: [ 0, 1, '', 'string', null, {}, [], , ]
+		i: [ 0, 1, "", "string", null, {}, [] ]
 	};
-	
-	it('create test data file', function() {
+
+	it("create test data file", function() {
 		fs.writeFileSync(testDataPath, JSON.stringify(testData));
 	});
 });
 
-describe('jsonWatcher', function() {
-	
+describe("jsonWatcher", function() {
 	var test = null;
-	
-	it('read', function(done) {
-		
-		chinachu.jsonWatcher(testDataPath, function(err, data, msg) {
-			should.strictEqual(null, err);
-			
+
+	it("read", function(done) {
+		var finished = false;
+
+		watcher = chinachu.jsonWatcher(testDataPath, function(err, data, msg) {
+			if (finished) {
+				return;
+			}
+
+			if (err) {
+				finished = true;
+				done(new Error(err));
+				return;
+			}
+
+			finished = true;
 			test = data;
-			
+
 			should.exist(test);
-			
+
 			done();
 		}, { now: true });
 	});
-	
-	it('validate', function() {
-		
+
+	it("validate", function() {
 		should.strictEqual(test.a, 0);
 		should.strictEqual(test.b, 1);
-		should.strictEqual(test.c, '');
-		should.strictEqual(test.d, 'string');
+		should.strictEqual(test.c, "");
+		should.strictEqual(test.d, "string");
 		should.strictEqual(test.e, null);
 	});
-	
-	it('watch');
+
+	it.skip("watch");
 });
 
-describe('(clean up)', function() {
-	
-	it('remove test data file', function() {
-		fs.unlinkSync(testDataPath);
+describe("(clean up)", function() {
+	it("remove test data file", function() {
+		if (watcher && typeof watcher.close === "function") {
+			watcher.close();
+		}
+
+		if (fs.existsSync(testDataPath)) {
+			fs.unlinkSync(testDataPath);
+		}
 	});
 });
